processes and initializations are moved to splash activity
