+++ CollectorAndroid/src/main/java/uk/ac/ucl/excites/sapelli/collector/CollectorApp.java
@@ -68,6 +68,7 @@ import uk.ac.ucl.excites.sapelli.shared.io.FileStorageException;
 import uk.ac.ucl.excites.sapelli.shared.util.TimeUtils;
 import uk.ac.ucl.excites.sapelli.shared.util.android.Debug;
 import uk.ac.ucl.excites.sapelli.shared.util.android.DeviceControl;
+import uk.ac.ucl.excites.sapelli.storage.StorageClient;
 import uk.ac.ucl.excites.sapelli.storage.db.RecordStore;
 import uk.ac.ucl.excites.sapelli.storage.db.sql.SQLRecordStoreUpgrader;
 import uk.ac.ucl.excites.sapelli.storage.db.sql.sqlite.SQLiteRecordStore;
@@ -80,15 +81,8 @@ import uk.ac.ucl.excites.sapelli.storage.db.sql.sqlite.android.AndroidSQLiteReco
  */
 public class CollectorApp extends Application
 {
-
-	// STATICS------------------------------------------------------------
-	static protected final String TAG = "CollectorApp";
-
-	static private final String DATABASE_BASENAME = "Sapelli";
 	static private final String DEMO_PREFIX = "Demo_";
-
-	static private final String CRASHLYTICS_VERSION_INFO = "VERSION_INFO";
-	static private final String CRASHLYTICS_BUILD_INFO = "BUILD_INFO";
+	static public final String DATABASE_BASENAME = "Sapelli";
 	static public final String CRASHLYTICS_DEVICE_ID_CRC32 = "SAPELLI_DEVICE_ID_CRC32";
 	static public final String CRASHLYTICS_DEVICE_ID_MD5 = "SAPELLI_DEVICE_ID_MD5";
 
@@ -97,288 +91,52 @@ public class CollectorApp extends Application
 	 */
 	static public final String PROPERTY_LAST_PROJECT = "SAPELLI_LAST_RUNNING_PROJECT";
 
-	public static enum StorageStatus
-	{
-		UNKNOWN, STORAGE_OK, STORAGE_UNAVAILABLE, STORAGE_REMOVED
-	}
-
 	// DYNAMICS-----------------------------------------------------------
 	private BuildInfo buildInfo;
-
+	public final AndroidCollectorClient collectorClient = new AndroidCollectorClient();
 	private CollectorPreferences preferences;
 
-	public final AndroidCollectorClient collectorClient = new AndroidCollectorClient();
 
 	// Files storage:
 	private FileStorageProvider fileStorageProvider;
 	private FileStorageException fileStorageException = null;
 
+
+	public BuildInfo getBuildInfo()
+	{
+		return buildInfo;
+	}
+
+	public void setBuildInfo(BuildInfo buildInfo) {
+		this.buildInfo = buildInfo;
+	}
+
+	public void setPreferences(CollectorPreferences preferences) {
+		this.preferences = preferences;
+	}
+
+	public CollectorPreferences getPreferences(){
+		return preferences;
+	}
+	public void setFileStorageProvider(FileStorageProvider fileStorageProvider) {
+		this.fileStorageProvider = fileStorageProvider;
+	}
+
 	@Override
-	public void onCreate()
+	public void onTerminate()
 	{
-		super.onCreate();
-
-		// Build info:
-		this.buildInfo = BuildInfo.GetInstance(getApplicationContext());
-
-		Debug.d("CollectorApp started.\nBuild info:\n" + buildInfo.getAllInfo());
-
-		// Start Fabric
-		setFabric();
-
-		// Set Timber for logging
-		setTimber();
-
-		// Set Stetho for debugging
-		setStetho();
-
-		// Get collector preferences:
-		preferences = new CollectorPreferences(getApplicationContext());
-
-		// Initialise file storage:
-		try
-		{
-			this.fileStorageProvider = initialiseFileStorage(); // throws FileStorageException
-		}
-		catch(FileStorageException fse)
-		{
-			this.fileStorageException = fse; // postpone throwing until getFileStorageProvider() is called!
-		}
-
-		// Set up a CrashReporter (will use dumps folder):
-		if(fileStorageProvider != null)
-			Thread.setDefaultUncaughtExceptionHandler(new CrashReporter(fileStorageProvider, getResources().getString(R.string.app_name)));
-
-		// Create shortcut to Sapelli Collector on Home Screen:
-		if(preferences.isFirstInstallation())
-		{
-			// Create shortcut
-			ProjectRunHelpers.createCollectorShortcut(getApplicationContext());
-			// Set first installation to false
-			preferences.setFirstInstallation(false);
-		}
+		super.onTerminate();
+		// This method is for use in emulated process environments. It will never be called on
+		// a production Android device, where processes are removed by simply killing them; no
+		// user code (including this callback) is executed when doing so.
+		Debug.d("Should never be called!");
 	}
 
-	/**
-	 * Set up Fabric
-	 */
-	private void setFabric()
+	@Override
+	public void onLowMemory()
 	{
-		// Set up Crashlytics, disabled for debug builds
-		final CrashlyticsCore crashlyticsCore = new CrashlyticsCore.Builder().disabled(BuildConfig.DEBUG).build();
-		final Crashlytics crashlyticsKit = new Crashlytics.Builder().core(crashlyticsCore).build();
-		Fabric.with(this, crashlyticsKit);
-
-		Crashlytics.setString(CRASHLYTICS_VERSION_INFO, buildInfo.getNameAndVersion() + " [" + buildInfo.getExtraVersionInfo() + "]");
-		Crashlytics.setString(CRASHLYTICS_BUILD_INFO, buildInfo.getBuildInfo());
-	}
-
-	/**
-	 * Set up Timber for logging
-	 */
-	private void setTimber()
-	{
-		// Enable Timber
-		if(BuildConfig.DEBUG)
-			Timber.plant(new DebugTree());
-		else
-			Timber.plant(new CrashlyticsTree());
-	}
-
-	/**
-	 * Set up Stetho for debugging
-	 */
-	private void setStetho()
-	{
-		// Enable Stetho in Debug versions
-		if(!BuildConfig.DEBUG)
-			return;
-
-		Timber.d("Enable Stetho");
-
-		Stetho.initialize(Stetho.newInitializerBuilder(this)
-		  .enableWebKitInspector(new InspectorModulesProvider()
-		  {
-			  @Override
-			  public Iterable<ChromeDevtoolsDomain> get()
-			  {
-				  return new Stetho.DefaultInspectorModulesBuilder(CollectorApp.this)
-					.provideDatabaseDriver(createCustomDatabaseDriver(CollectorApp.this))
-					.finish();
-			  }
-		  }).build());
-	}
-
-	private SqliteDatabaseDriver createCustomDatabaseDriver(Context context)
-	{
-		return new SqliteDatabaseDriver(context, new DatabaseFilesProvider()
-		{
-			@Override
-			public List<File> getDatabaseFiles()
-			{
-				List<File> dbs = new ArrayList<>();
-				final String dbPath = SQLiteRecordStore.GetDBFileName(fileStorageProvider.getDBFolder(false).getAbsolutePath() + File.separator + DATABASE_BASENAME);
-				Timber.d("Try to connect to db at: %s", dbPath);
-				dbs.add(new File(dbPath));
-				return dbs;
-			}
-		}, new DefaultDatabaseConnectionProvider());
-	}
-
-	/**
-	 * @return
-	 * @throws FileStorageException
-	 */
-	private FileStorageProvider initialiseFileStorage() throws FileStorageException
-	{
-		File sapelliFolder = null;
-
-		// Try to get Sapelli folder path from preferences:
-		try
-		{
-			sapelliFolder = new File(preferences.getSapelliFolderPath());
-		}
-		catch(NullPointerException npe)
-		{
-		}
-
-		// Did we get the folder path from preferences? ...
-		if(sapelliFolder == null)
-		{    // No: first installation or reset
-
-			// Find appropriate files dir (using application-specific folder, which is removed upon app uninstall!):
-			File[] paths = DeviceControl.getExternalFilesDirs(this, null);
-			if(paths != null && paths.length != 0)
-			{
-				// We count backwards because we prefer secondary external storage (which is likely to be on an SD card rather unremovable memory)
-				for(int p = paths.length - 1; p >= 0; p--)
-					if(isMountedReadableWritableDir(paths[p]))
-					{
-						sapelliFolder = paths[p];
-						break;
-					}
-			}
-
-			// Do we have a path?
-			if(sapelliFolder != null)
-				// Yes: store it in the preferences:
-				preferences.setSapelliFolder(sapelliFolder.getAbsolutePath());
-			else
-				// No :-(
-				throw new FileStorageUnavailableException();
-		}
-		else
-		{    // Yes, we got path from preferences, check if it is available ...
-			if(!isMountedReadableWritableDir(sapelliFolder)) // (will also attempt to create the directory if it doesn't exist)
-				// No :-(
-				throw new FileStorageRemovedException(sapelliFolder.getAbsolutePath());
-		}
-
-		// If we get here this means we have a non-null sapelliFolder object representing an accessible path...
-
-		// Try to get the Android Downloads folder...
-		File downloadsFolder = Environment.getExternalStoragePublicDirectory(Environment.DIRECTORY_DOWNLOADS);
-		if(!isMountedReadableWritableDir(downloadsFolder)) // check if we can access it (will also attempt to create the directory if it doesn't exist)
-			// No :-(
-			throw new FileStorageException("Cannot access downloads folder: " + downloadsFolder.getAbsolutePath());
-
-		// Create a test database to get the DB folder
-		TestDatabaseHelper testDatabaseHelper = new TestDatabaseHelper(this);
-
-		// Try to get the Databases folder...
-		File databaseFolder = testDatabaseHelper.getDatabaseFolder();
-		if(!isMountedReadableWritableDir(databaseFolder)) // check if we can access it (will also attempt to create the directory if it doesn't exist)
-			// No :-(
-			throw new FileStorageException("Cannot access database folder: " + databaseFolder.getAbsolutePath());
-
-		final AndroidFileStorageProvider androidFileStorageProvider = new AndroidFileStorageProvider(sapelliFolder, databaseFolder, downloadsFolder);
-
-		moveDB(androidFileStorageProvider);
-
-		return androidFileStorageProvider; // Android specific subclass of FileStorageProvider, which generates .nomedia files
-	}
-
-	/**
-	 * The location for the Sapelli database has changed on Sapelli v2.0.0 beta 27, therefore move
-	 * the DB from the old location to the new one.
-	 *
-	 * @param androidFileStorageProvider
-	 */
-	private void moveDB(AndroidFileStorageProvider androidFileStorageProvider)
-	{
-
-		final File oldDBFolder = androidFileStorageProvider.getOldDBFolder(false);
-		final File nedDBFolder = androidFileStorageProvider.getDBFolder(false);
-		Timber.d("Old DB path: %s", oldDBFolder);
-		Timber.d("New DB path: %s", nedDBFolder);
-
-		try
-		{
-			File oldDB = new File(SQLiteRecordStore.GetDBFileName(oldDBFolder.getAbsolutePath() + File.separator + DATABASE_BASENAME));
-			File newDB = new File(nedDBFolder + File.separator + oldDB.getName());
-
-			if(!newDB.exists())
-				newDB.createNewFile();
-
-			if(oldDB.exists() && newDB.exists())
-			{
-				Timber.d("Move Old DB: %s to %s", oldDB, newDB);
-
-				FileChannel src = new FileInputStream(oldDB).getChannel();
-				FileChannel dst = new FileOutputStream(newDB).getChannel();
-				dst.transferFrom(src, 0, src.size());
-				src.close();
-				dst.close();
-
-				// Delete the old DB
-				FileUtils.deleteQuietly(oldDB);
-
-				// Delete all other files in the old DB e.g. the journal etc.
-				for(File file : oldDBFolder.listFiles())
-					FileUtils.deleteQuietly(file);
-
-				// Finally delete the old directory
-				FileUtils.deleteQuietly(oldDBFolder);
-			}
-		}
-		catch(Exception e)
-		{
-			Timber.e(e);
-		}
-	}
-
-	/**
-	 * Create a "Test.db" in the default location of Android. Use this to get the directory where
-	 * Android stores by default the SQLite database
-	 */
-	public class TestDatabaseHelper extends SQLiteOpenHelper
-	{
-		private static final String DATABASE_NAME = "Test.db";
-		private static final int DATABASE_VERSION = 1;
-		private Context context;
-
-		public TestDatabaseHelper(Context context)
-		{
-			super(context, DATABASE_NAME, null, DATABASE_VERSION);
-			this.context = context;
-		}
-
-		@Override
-		public void onCreate(SQLiteDatabase sqLiteDatabase)
-		{
-			// Do nothing
-		}
-
-		@Override
-		public void onUpgrade(SQLiteDatabase sqLiteDatabase, int i, int i1)
-		{
-			// Do nothing
-		}
-
-		public File getDatabaseFolder()
-		{
-			return context.getDatabasePath(DATABASE_NAME).getParentFile();
-		}
+		super.onLowMemory();
+		Debug.d("onLowMemory() called!");
 	}
 
 	/**
@@ -398,51 +156,13 @@ public class CollectorApp extends Application
 	}
 
 	/**
-	 * @return the preferences
+	 * @return handles for all Stores that need to be backed-up
 	 */
-	public CollectorPreferences getPreferences()
+	public StoreHandle<?>[] getStoreHandlesForBackup()
 	{
-		return preferences;
-	}
-
-	/**
-	 * Check if a directory is on a mounted storage and writable/readable
-	 *
-	 * @param dir
-	 * @return
-	 * @throws FileStorageException
-	 */
-	private boolean isMountedReadableWritableDir(File dir) throws FileStorageException
-	{
-		try
-		{
-			return    // Null check:
-			  (dir != null)
-				// Try to create the directory if it is not there
-				&& FileHelpers.createDirectory(dir)
-					/* Check storage state, accepting both MEDIA_MOUNTED and MEDIA_UNKNOWN.
-					 * 	The MEDIA_UNKNOWN state occurs when a path isn't backed by known storage media; e.g. the SD Card on
-					 * the Samsung Xcover 2 (the detection of which we have to force in DeviceControl#getExternalFilesDirs()). */
-				&& (Environment.MEDIA_MOUNTED.equals(EnvironmentCompat.getStorageState(dir)) || EnvironmentCompat.MEDIA_UNKNOWN.equals(EnvironmentCompat.getStorageState(dir)))
-				// Check whether we have read & write access to the directory:
-				&& FileHelpers.isReadableWritableDirectory(dir);
-		}
-		catch(Exception e)
-		{
-			throw new FileStorageException("Unable to create or determine status of directory: " + (dir != null ? dir.getAbsolutePath() : "null"), e);
-		}
-	}
-
-	@Override
-	public void onConfigurationChanged(Configuration newConfig)
-	{
-		super.onConfigurationChanged(newConfig);
-		// Debug.d(newConfig.toString());
-	}
-
-	public BuildInfo getBuildInfo()
-	{
-		return buildInfo;
+		return new StoreHandle[]{collectorClient.recordStoreHandle,
+				collectorClient.transmissionStoreHandle,
+				collectorClient.projectStoreHandle};
 	}
 
 	/**
@@ -455,33 +175,6 @@ public class CollectorApp extends Application
 	public String getDemoPrefix()
 	{
 		return (buildInfo.isDemoBuild() ? DEMO_PREFIX + FileHelpers.makeValidFileName(TimeUtils.getTimestampForFileName(buildInfo.getTimeStamp())) : "");
-	}
-
-	@Override
-	public void onLowMemory()
-	{
-		super.onLowMemory();
-		Debug.d("onLowMemory() called!");
-	}
-
-	@Override
-	public void onTerminate()
-	{
-		super.onTerminate();
-		// This method is for use in emulated process environments. It will never be called on
-		// a production Android device, where processes are removed by simply killing them; no
-		// user code (including this callback) is executed when doing so.
-		Debug.d("Should never be called!");
-	}
-
-	/**
-	 * @return handles for all Stores that need to be backed-up
-	 */
-	public StoreHandle<?>[] getStoreHandlesForBackup()
-	{
-		return new StoreHandle[]{collectorClient.recordStoreHandle,
-		  collectorClient.transmissionStoreHandle,
-		  collectorClient.projectStoreHandle};
 	}
 
 	/**
@@ -500,7 +193,7 @@ public class CollectorApp extends Application
 		}
 
 		@Override
-		protected void createAndSetRecordStore(StoreSetter<RecordStore> setter) throws DBException
+		protected void createAndSetRecordStore(StoreHandle.StoreSetter<RecordStore> setter) throws DBException
 		{
 			@SuppressWarnings("resource")
 			RecordStore recordStore = new AndroidSQLiteRecordStore(this, CollectorApp.this, getFileStorageProvider().getDBFolder(true), getDemoPrefix() /*will be "" if not in demo mode*/ + DATABASE_BASENAME, CURRENT_COLLECTOR_RECORDSTORE_VERSION, new CollectorSQLRecordStoreUpgrader(this, this, getFileStorageProvider()));
@@ -513,7 +206,7 @@ public class CollectorApp extends Application
 		}
 
 		@Override
-		protected void createAndSetProjectStore(StoreSetter<ProjectStore> setter) throws DBException
+		protected void createAndSetProjectStore(StoreHandle.StoreSetter<ProjectStore> setter) throws DBException
 		{
 			setter.setAndInitialise(new ProjectRecordStore(this, getFileStorageProvider()));
 			//setter.setAndInitialise(new PrefProjectStore(CollectorApp.this, getFileStorageProvider(), getDemoPrefix()));
@@ -576,5 +269,4 @@ public class CollectorApp extends Application
 		}
 
 	}
-
 }
