+++ CollectorAndroid/src/main/res/values/strings.xml
@@ -1,5 +1,4 @@
-<?xml version="1.0" encoding="utf-8" standalone="no"?>
-<!--
+<?xml version="1.0" encoding="utf-8" standalone="no"?><!--
 	Sapelli data collection platform: http://sapelli.org
 	
 	Copyright 2012-2016 University College London - ExCiteS group
@@ -16,8 +15,7 @@
 	See the License for the specific language governing permissions and
 	limitations under the License.
  -->
-<resources xmlns:tools="http://schemas.android.com/tools"
-           tools:ignore="ExtraTranslation" tools:locale="en">
+<resources xmlns:tools="http://schemas.android.com/tools" tools:ignore="ExtraTranslation" tools:locale="en">
     <string name="app_name">Sapelli Collector</string>
     <string name="sapelli">Sapelli</string>
     <string name="collector">Collector</string>
